+++ src/main/java/book/store/service/category/CategoryServiceImpl.java
@@ -1,11 +1,12 @@
 package book.store.service.category;
 
-import book.store.dto.CategoryUpdateDto;
 import book.store.dto.category.CategoryResponseDto;
+import book.store.dto.category.CategoryUpdateDto;
 import book.store.dto.category.CreateCategoryRequestDto;
 import book.store.mapper.CategoryMapper;
 import book.store.model.Category;
 import book.store.repository.CategoryRepository;
+import book.store.telegram.strategy.notification.AdminNotificationService;
 import jakarta.persistence.EntityNotFoundException;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -16,13 +17,19 @@ import org.springframework.stereotype.Service;
 @RequiredArgsConstructor
 @Service
 public class CategoryServiceImpl implements CategoryService {
+    private static final String TELEGRAM = "Telegram";
+    private static final String CATEGORY_CREATION = "Category creation";
+    private static final String CATEGORY_UPDATING = "Category updating";
+    private static final String CATEGORY_DELETING = "Category deleting";
     private final CategoryRepository categoryRepository;
     private final CategoryMapper categoryMapper;
+    private final List<AdminNotificationService<Category>> notificationServices;
 
     @Override
     public CategoryResponseDto create(CreateCategoryRequestDto requestDto) {
         Category category = categoryMapper.toModel(requestDto);
         categoryRepository.save(category);
+        sendMessage(TELEGRAM, CATEGORY_CREATION, null, category);
         return categoryMapper.toResponseDto(category);
     }
 
@@ -44,7 +51,11 @@ public class CategoryServiceImpl implements CategoryService {
 
     @Override
     public void deleteById(Long id) {
+        if (categoryRepository.findById(id).isEmpty()) {
+            return;
+        }
         categoryRepository.deleteById(id);
+        sendMessage(TELEGRAM, CATEGORY_DELETING, null, new Category(id));
     }
 
     @Override
@@ -54,6 +65,21 @@ public class CategoryServiceImpl implements CategoryService {
                         "Can't find a category by id " + id));
         category = categoryMapper.toModel(category, updateDto);
         categoryRepository.save(category);
+        sendMessage(TELEGRAM, CATEGORY_UPDATING, null, category);
         return categoryMapper.toResponseDto(category);
+    }
+
+    private void sendMessage(
+            String notificationService,
+            String messageType,
+            Long chatId,
+            Category book) {
+        notificationServices
+                .stream()
+                .filter(service -> service.isApplicable(notificationService, messageType))
+                .findFirst()
+                .orElseThrow(() -> new RuntimeException(
+                        "Can't find a notification service for " + messageType))
+                .sendMessage(chatId, book);
     }
 }
